+++ src/pages/lobby/update.ts
@@ -1,9 +1,14 @@
 import { setupToolbar } from "./toolbar";
 import { setupUserList } from "./userList";
 import { setupChat } from "./chat";
-import { Direction, NavigationProps } from "@app/types/navigation";
-import { findFirstFocusableChild } from "@app/dom";
-import { dispatchOOBEvent } from "@app/oobNavigator";
+import {
+  onChatHorizontalOOB,
+  onChatVerticalOOB,
+  onOOBNavigation,
+  onToolbarVerticalOOB,
+  onUserListHorizontalOOB,
+  onUserListVerticalOOB,
+} from "@app/pages/lobby/navigation";
 
 const teardown: Teardown[] = [];
 export function updateLobby(root: HTMLElement) {
@@ -39,9 +44,10 @@ export function updateLobby(root: HTMLElement) {
     setupToolbar({
       root,
       onOOBNavigation: onOOBNavigation.bind(null, root),
-      onVerticalOOB: onToolbarOOB.bind(null, usersListRoot),
+      onVerticalOOB: onToolbarVerticalOOB.bind(null, usersListRoot),
     })
   );
+
   teardown.push(
     setupChat({
       root: chatRoot,
@@ -50,57 +56,4 @@ export function updateLobby(root: HTMLElement) {
       onOOBNavigation: onOOBNavigation.bind(null, root),
     })
   );
-}
-
-function onOOBNavigation(root: HTMLElement, props: NavigationProps) {
-  dispatchOOBEvent({
-    ...props,
-    root,
-  });
-}
-
-function onToolbarOOB(usersListRoot: HTMLElement, direction: Direction) {
-  if (direction === "END") {
-    const next = findFirstFocusableChild(usersListRoot);
-    next?.focus();
-  }
-}
-
-function onUserListVerticalOOB(toolbarRoot: HTMLElement, direction: Direction) {
-  if (direction === "START") {
-    const next = findFirstFocusableChild(toolbarRoot);
-    next?.focus();
-  }
-}
-
-function onUserListHorizontalOOB(chatRoot: HTMLElement, direction: Direction) {
-  if (direction === "START") {
-    const next = findFirstFocusableChild(chatRoot);
-    next?.focus();
-  }
-}
-
-function onChatHorizontalOOB(
-  root: HTMLElement,
-  userList: HTMLElement,
-  direction: Direction,
-  el: HTMLElement
-) {
-  if (direction === "END") {
-    const next = findFirstFocusableChild(userList);
-    next?.focus();
-  } else if (direction === "START") {
-    // Dispatch to
-    onOOBNavigation(root, {
-      axis: "HORIZONTAL",
-      direction,
-      el,
-    });
-  }
-}
-function onChatVerticalOOB(toolbarRoot: HTMLElement, direction: Direction) {
-  if (direction === "START") {
-    const next = findFirstFocusableChild(toolbarRoot);
-    next?.focus();
-  }
 }
