+++ testdata/lobby-has-been-challenged/index.html
@@ -1019,8 +1019,6 @@
     </div> <!----></div> <div data-v-42a8da26="" class="rankWrapper"><!----></div> <div data-v-42a8da26="" class="pingWrapper"><!----></div></div></div></div> <!----> <!----> <!----></div></div></div></div> <!----> <!----></div><script type="text/javascript" src="/static/js/manifest.c22a0c3defdf1eb18d28.js"></script><script type="text/javascript" src="/static/js/vendor.b794f10df6592d5dec53.js"></script><script type="text/javascript" src="/static/js/app.c46c3a02a4f21f51373d.js"></script><script src="https://www.google.com/recaptcha/api.js?onload=vueRecaptchaInit&amp;render=explicit" async="" defer=""></script>
  
     <!-- Added by Me -->
-    <!-- 
     <script src="/out/inject.js"></script>
--->
 
    </body></html>
